% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chckapmalawi.R
\docType{data}
\name{chckapmalawi}
\alias{chckapmalawi}
\title{Community Health Club Knowledge Attitude Practice Survey Malawi}
\format{
A tibble with 500 rows and 4 variables:
\describe{
\item{village}{ Village in Malawi.}
\item{gender}{ Gender of the respondent.}
\item{age}{ Age of the respondent.}
\item{water_pay}{ Question: Do you believe you should pay for your water?}
}
}
\usage{
chckapmalawi
}
\description{
A survey conducted by BASEflow for Waste Advisers, Malawi. The survey was
implemented in 20XX.
}
\keyword{datasets}
